+++ filament/backend/src/noop/NoopDriver.cpp
@@ -165,7 +165,7 @@ bool NoopDriver::isFrameTimeSupported() {
 }
 
 math::float2 NoopDriver::getClipSpaceParams() {
-    return math::float2{ -1.0f, 0.0f };
+    return math::float2{ 1.0f, 0.0f };
 }
 
 uint8_t NoopDriver::getMaxDrawBuffers() {