+++ RELEASE_NOTES.md
@@ -3,7 +3,15 @@
 This file contains one line summaries of commits that are worthy of mentioning in release notes.
 A new header is inserted each time a *tag* is created.
 
-## v1.12.1 (currently main branch)
+## v1.12.2 (currently main branch)
+
+## v1.12.1
+
+- engine: `double` precision translation support in TransformManager. Disabled by default.
+  Augment model (and view) matrix on `Camera` to accept double precision matrices. When enabled,
+  double precision translations allow filament to handle a very large world space.
+  [**New API**].
+- engine: Fix, Views with custom render targets are now blendable.
 
 ## v1.12.0